Maximum Likelhood Analysis

Maximum Likelhood Analysis

ml_analyse(expr, pars, data)

Arguments

expr
A string of R code that evaluates to the log-likelihood.

pars
An named list of the parameter initial values (coerced to an nlist object).

data
An named list of the data (coerced to an nlist object).

Value

An ml_analysis object.

Examples

template <- "sum(dnorm(len, mu, sigma, log = TRUE))"
pars <- list(mu = 20, sigma = 8)
data <- datasets::ToothGrowth
analysis <- ml_analyse(template, pars = pars, data = data)
analysis

#> $expr
#> [1] "sum(dnorm(len, mu, sigma, log = TRUE))"
#> 
#> $pars
#> $mu
#> [1] 20
#> 
#> $sigma
#> [1] 8
#> 
#> an nlist object with 2 natomic elements
#> $data
#>     len supp dose
#> 1   4.2   VC  0.5
#> 2  11.5   VC  0.5
#> 3   7.3   VC  0.5
#> 4   5.8   VC  0.5
#> 5   6.4   VC  0.5
#> 6  10.0   VC  0.5
#> 7  11.2   VC  0.5
#> 8  11.2   VC  0.5
#> 9   5.2   VC  0.5
#> 10  7.0   VC  0.5
#> 11 16.5   VC  1.0
#> 12 16.5   VC  1.0
#> 13 15.2   VC  1.0
#> 14 17.3   VC  1.0
#> 15 22.5   VC  1.0
#> 16 17.3   VC  1.0
#> 17 13.6   VC  1.0
#> 18 14.5   VC  1.0
#> 19 18.8   VC  1.0
#> 20 15.5   VC  1.0
#> 21 23.6   VC  2.0
#> 22 18.5   VC  2.0
#> 23 33.9   VC  2.0
#> 24 25.5   VC  2.0
#> 25 26.4   VC  2.0
#> 26 32.5   VC  2.0
#> 27 26.7   VC  2.0
#> 28 21.5   VC  2.0
#> 29 23.3   VC  2.0
#> 30 29.5   VC  2.0
#> 31 15.2   OJ  0.5
#> 32 21.5   OJ  0.5
#> 33 17.6   OJ  0.5
#> 34  9.7   OJ  0.5
#> 35 14.5   OJ  0.5
#> 36 10.0   OJ  0.5
#> 37  8.2   OJ  0.5
#> 38  9.4   OJ  0.5
#> 39 16.5   OJ  0.5
#> 40  9.7   OJ  0.5
#> 41 19.7   OJ  1.0
#> 42 23.3   OJ  1.0
#> 43 23.6   OJ  1.0
#> 44 26.4   OJ  1.0
#> 45 20.0   OJ  1.0
#> 46 25.2   OJ  1.0
#> 47 25.8   OJ  1.0
#> 48 21.2   OJ  1.0
#> 49 14.5   OJ  1.0
#> 50 27.3   OJ  1.0
#> 51 25.5   OJ  2.0
#> 52 26.4   OJ  2.0
#> 53 22.4   OJ  2.0
#> 54 24.5   OJ  2.0
#> 55 24.8   OJ  2.0
#> 56 30.9   OJ  2.0
#> 57 26.4   OJ  2.0
#> 58 27.3   OJ  2.0
#> 59 29.4   OJ  2.0
#> 60 23.0   OJ  2.0
#> 
#> $optim
#> $optim$par
#>        mu     sigma 
#> 18.813413  7.585305 
#> 
#> $optim$value
#> [1] -206.7091
#> 
#> $optim$counts
#> function gradient 
#>       16        5 
#> 
#> $optim$convergence
#> [1] 0
#> 
#> $optim$message
#> NULL
#> 
#> $optim$hessian
#>                  mu         sigma
#> mu    -1.042810e+00  2.182432e-05
#> sigma  2.182432e-05 -2.085619e+00
#> 
#> 
#> attr(,"class")
#> [1] "ml_analysis"